# Optimize string concatenation