--- a/benches/convert_bench.rs
+++ b/benches/convert_bench.rs
@@ -15,7 +15,7 @@ fn benchmark(c: &mut Criterion) {
 
 criterion_group!(
     name = benches;
-    config = Criterion::default().measurement_time(Duration::from_secs(12));
+    config = Criterion::default().measurement_time(Duration::from_secs(20));
     targets = benchmark
 );
 criterion_main!(benches);