--- a/src/element_handler/blockquote.rs
+++ b/src/element_handler/blockquote.rs
@@ -1,12 +1,15 @@
-use crate::{text_util::TrimAsciiWhitespace, Element};
+use crate::{
+    text_util::{concat_strings, TrimAsciiWhitespace},
+    Element,
+};
 
 pub(super) fn blockquote_handler(element: Element) -> Option<String> {
     let content = element.content.trim_start_matches(|ch| ch == '\n');
     let content = content
         .trim_end_ascii_whitespace()
         .lines()
-        .map(|line| format!("> {}", line))
+        .map(|line| concat_strings!("> ", line))
         .collect::<Vec<String>>()
         .join("\n");
-    Some(format!("\n\n{}\n\n", content))
+    Some(concat_strings!("\n\n", content, "\n\n"))
 }