--- a/src/element_handler/blockquote.rs
+++ b/src/element_handler/blockquote.rs
@@ -1,5 +1,5 @@
 use crate::{
-    text_util::{concat_strings, TrimAsciiWhitespace},
+    text_util::{concat_strings, JoinOnStringIterator, TrimAsciiWhitespace},
     Element,
 };
 
@@ -9,7 +9,6 @@ pub(super) fn blockquote_handler(element: Element) -> Option<String> {
         .trim_end_ascii_whitespace()
         .lines()
         .map(|line| concat_strings!("> ", line))
-        .collect::<Vec<String>>()
         .join("\n");
     Some(concat_strings!("\n\n", content, "\n\n"))
 }