~20% faster in the built-in benchmark!

```diff
- convert()               time:   [107.38 ms 108.01 ms 108.68 ms]
+ convert()               time:   [88.528 ms 89.545 ms 90.670 ms]
```